cli: don't inherit DvcParserError from DvcException, move to dvc.cli
